app/schemas/transaction.py
from pydantic import BaseModel
from datetime import datetime

class TransactionCreate(BaseModel):
    amount: float
    description: str

class TransactionOut(TransactionCreate):
    id: int
    timestamp: datetime

    class Config:
        orm_mode = True
